pyproject.toml
[build-system]
requires = ["setuptools==68.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skelet"
version = "0.0.6"
authors = [{ name = "Evgeniy Blinov", email = "zheni-b@yandex.ru" }]
description = 'Collect all the settings in one place'
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    'printo>=0.0.10',
    'locklib>=0.0.20',
    'simtypes>=0.0.10',
    'tomli==2.3.0 ; python_version <= "3.12"',
    'pyyaml==6.0.3',
    'types-pyyaml==6.0.12.20241230',
    'denial>=0.0.11',
    'sigmatch>=0.0.6',
]
classifiers = [
    "Operating System :: OS Independent",
    'Operating System :: MacOS :: MacOS X',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: POSIX',
    'Operating System :: POSIX :: Linux',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3.14',
    'Programming Language :: Python :: Free Threading',
    'Programming Language :: Python :: Free Threading :: 3 - Stable',
    'License :: OSI Approved :: MIT License',
    'Intended Audience :: Developers',
    'Topic :: Software Development :: Libraries',
]
keywords = ['settings', 'configs']

[tool.setuptools.package-data]
"skelet" = ["py.typed"]

[tool.mutmut]
paths_to_mutate = "skelet"
runner = "pytest"

[tool.pytest.ini_options]
markers = ["mypy_testing"]

[tool.ruff]
lint.ignore = ['E501', 'E712', 'PTH123', 'PTH118', 'PLR2004', 'PTH107', 'SIM105', 'SIM102', 'RET503', 'PLR0912', 'C901']
lint.select = ["ERA001", "YTT", "ASYNC", "BLE", "B", "A", "COM", "INP", "PIE", "T20", "PT", "RSE", "RET", "SIM", "SLOT", "TID252", "ARG", "PTH", "I", "C90", "N", "E", "W", "D201", "D202", "D419", "F", "PL", "PLE", "PLR", "PLW", "RUF", "TRY201", "TRY400", "TRY401"]
format.quote-style = "single"

[project.urls]
'Source' = 'https://github.com/pomponchik/skelet'
'Tracker' = 'https://github.com/pomponchik/skelet/issues'
